provides full implementation for search and favourites (no brewery details yet)
this version provides a working system for the search and the favourite
page. this is handled by the Brewery component which for now has a star
that identify the brewery as favourite or not.

MobX store is now also holding the information regarding the current
search array: this way when i go back and forth between pages i will
always see the last search loaded on screen