BUGFIX ts and mapEvents
seems that the compiler and ts are struggling with the newest version of
the js google map api, so i had to do a workaround to ake sure that the eventListener for the zoom_changed is correctly registered. to do so, i create a variable to store it outside the check for coordinates (hence it's created wheter i have the coordinates or not), this way the cleaning method will always remove it from the page on unmount, so it's not an issue.