adds clearbit logo api
the usage was easier than expected: it just needed to call
https://logo.clearbit.com/${url} where url is the website of the brewery
and it will automatically scrape the image from the website. by doing so
it was possible to add the url as above directly on the img src
attribute, then manage the onError or the lack of website_url. onLoad it
will also take care to show a loading state on screen while the image is
being fetched