TESTING: render.test now working correctly
this happened as the basic <App /> component won't run in the test without being rebuilt with a memoryRouter pointing to a / route. this allowed for the /route to be present and so show the searchbox as well. previously this mechanism was not understood and it seemed like this part was workign as intended